refact: solve circular import
